requirements.txt
cyhunspell==2.0.2
gensim==4.1.2
nltk>=3.7
numpy>=1.22.2
pandas>=1.4.1
pyyaml>=5.4.1
scipy>=1.8.0
